Haskell: refactor
Use pointfree style more, move some functions to the top level, and
rewrite some functions to be more concise. Additionally, misuse of `or`
to mean `uncurry (||)` on a `(Bool, Bool)` has been corrected (this
was silently failing and acting as `snd` because it interpreted `(Bool,
Bool)` as a container of `Bool` by seeing it as `((,) Bool) Bool`).